src/test/kotlin/dev/shtanko/algorithms/leetcode/MinimumCostToConnectSticksTest.kt
package dev.shtanko.algorithms.leetcode

import org.hamcrest.CoreMatchers.equalTo
import org.hamcrest.MatcherAssert.assertThat
import org.junit.jupiter.api.extension.ExtensionContext
import org.junit.jupiter.params.ParameterizedTest
import org.junit.jupiter.params.provider.Arguments
import org.junit.jupiter.params.provider.ArgumentsProvider
import org.junit.jupiter.params.provider.ArgumentsSource
import java.util.stream.Stream

internal class MinimumCostToConnectSticksTest {

    internal class InputArgumentsProvider : ArgumentsProvider {
        override fun provideArguments(context: ExtensionContext?): Stream<out Arguments> = Stream.of(
            Arguments.of(intArrayOf(2, 4, 3), 14),
            Arguments.of(intArrayOf(1, 8, 3, 5), 30),
        )
    }

    @ParameterizedTest
    @ArgumentsSource(InputArgumentsProvider::class)
    internal fun `connect sticks test`(sticks: IntArray, expected: Int) {
        val actual = MinimumCostToConnectSticks().connectSticks(sticks)
        assertThat(actual, equalTo(expected))
    }
}
